% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distedge.R
\name{distedge}
\alias{distedge}
\title{Calculate raster distance to an edge.}
\usage{
distedge(DT, rowvar = "hdfrow", colvar = "hdfcol",
  nonmissingmask = "revna")
}
\arguments{
\item{DT}{a data.table}

\item{rowvar}{character of the column name for row}

\item{colvar}{character of the column name for column}

\item{nonmissingmask}{character of the column name for the non missing mask}
}
\value{
vector of distance to edge.
}
\description{
Calculate raster distance to an edge.
}
\examples{
 
aod20150211 <- merge(aod20150211, basegrid, by = "aodid", all.x = TRUE)
aod20150211[, revna := NA_integer_]
aod20150211[is.na(aod47), revna := 1]
distedge(aod20150211, rowvar = "hdfrow", colvar = "hdfcol", nonmissingmask = "revna")
}

