% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hdftoraster.R
\name{hdftoraster}
\alias{hdftoraster}
\title{Import an hdf file into a RasterLayer object.}
\usage{
hdftoraster(path, n)
}
\arguments{
\item{path}{path of the hdf file}

\item{n}{number of the file the user wants to import}
}
\value{
a RasterLayer object.
}
\description{
Import an hdf file into a RasterLayer object.
}

