% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfit.R
\name{modelfit}
\alias{modelfit}
\title{Summarize linear model fit.}
\usage{
modelfit(dataframe, yvar = "mean_pm25", xvar = "aod47")
}
\arguments{
\item{dataframe}{a data.frame}

\item{yvar}{character vector of the column name for outcome}

\item{xvar}{character vector of the column name(s) for predictor(s)}
}
\value{
a data.frame of the r-squared and the rmse for the linear regression
}
\description{
Summarize linear model fit.
}

