% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestbyday.R
\name{nearestbyday}
\alias{nearestbyday}
\title{Merge two point datasets by day.}
\usage{
nearestbyday(matrix1, matrix2, dt1, dt2, dt1varname, dt2varname,
  joinprefix = "closest", varstoget = "avewsp", knearest = 5,
  maxdistance = NA, nearestmean = FALSE)
}
\arguments{
\item{matrix1}{}

\item{matrix2, }{}

\item{dt1, }{}

\item{dt2, }{}

\item{dt1varname}{}

\item{joinprefix}{string to use as name for new variables}

\item{knearest}{number of nearest neighbors to check for non-missing data}

\item{maxdistance}{(optional) constraint on distance}

\item{nearestmean}{(optional) compute mean of knearest values within maxdistance?}
}
\value{
A data.table, akin to a semi_join on nearest neighbor and day.
}
\description{
Merge point datasets by day to the closest of K nearest neighbors 
 (with an optional distance constraint).
}
\examples{
jointo <- data.table(x = 1:5, y = 1:5, siteidx = 1:5,
                    day = as.Date("2004-04-01"))
joinfrom <- data.table(x = rep(1:5 + 0.1, times = 5),
                      y = rep(1:5, each = 5), 
                      siteidy = letters[1:25], # currently this needs to be character
                      value = 100 + 1:25,
                      day = as.Date("2004-04-01"))
                      
jointo.pt <- makepointsmatrix(datatable = jointo, 
                             xvar = "x", yvar = "y", idvar = "siteidx") 
joinfrom.pt <- makepointsmatrix(datatable = joinfrom, 
                               xvar = "x", yvar = "y", idvar = "siteidy") 
joinout <- nearestbyday(matrix1 = jointo.pt, matrix2 = joinfrom.pt, 
                       dt1 = jointo, dt2 = joinfrom, 
                       dt1varname = "siteidx", dt2varname = "siteidy", 
                       joinprefix = "nearest", varstoget = "value", 
                       knearest = 3, maxdistance = 2, nearestmean = TRUE)
}

