% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotraster.R
\name{plotraster}
\alias{plotraster}
\title{Plot a semi-regular grid as a ggmap.}
\usage{
plotraster(DT, basegrid, rowvar = "hdfrow", colvar = "hdfcol",
  longvar = "lon", latvar = "lat", id, zvar, ggmap = TRUE, zoom = 9,
  loncen = NA, latcen = NA, alpha.layer = 0.5, alpha.white.backg = 0.5,
  aspratio = 1.05)
}
\arguments{
\item{DT}{data.table for a single scene to be plotted}

\item{basegrid}{data.table that contains the row, col, lon and lat of the grid}

\item{rowvar}{character of the row name in basegrid}

\item{colvar}{character of the column name in basegrid}

\item{longvar}{character of the longitude name in basegrid}

\item{latvar}{character of latitude name in basegrid}

\item{id}{character of the ID of each cell of the grid - it has to be present both in DT and basegrid}

\item{zvar}{character of the variable in DT the user wants to plot}

\item{ggmap}{if TRUE (default) print the ggmap}

\item{zoom}{integer for the zoom of the background map}

\item{loncen}{numeric value for the longitude of the center of the background map}

\item{latcen}{numeric value for the latitude of the center of the background map}

\item{alpha.layer}{numeric value for the alpha of the layer}

\item{alpha.white.backg}{numeric value for the alpha of the background map}

\item{aspratio}{numeric value for the aspect ratio of the background map}
}
\value{
a ggmap object.
}
\description{
Plot a semi-regular grid as a ggmap.
}
\examples{
plotraster(aod20150211, basegrid, id = "aodid", zvar = "aod47")
}

