% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makepointsmatrix.R
\name{makepointsmatrix}
\alias{makepointsmatrix}
\title{Construct points matrix.}
\usage{
makepointsmatrix(datatable, xvar, yvar, idvar)
}
\arguments{
\item{datatable}{a data.table}

\item{xvar}{column name for x-coordinate}

\item{yvar}{column name for y-coordinate}

\item{idvar}{column name for unique key; currently better to use 
character (gets stored as row.names)}
}
\value{
a matrix of unique x, y pairs with rownames
}
\description{
Construct points matrix.
}

