% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelfit.R
\name{modelfit}
\alias{modelfit}
\title{Summarize linear model fit.}
\usage{
modelfit(DF, yvar = "pm", xvar = "aod47", verbose = TRUE, digits = 2)
}
\arguments{
\item{DF}{a data.frame}

\item{yvar}{character vector of the column name for outcome}

\item{xvar}{character vector of the column name(s) for predictor(s)}

\item{verbose}{logical whether to print summary}

\item{digits}{integer number of digits to print (if verbose is equal to FALSE, this parameter is not considered)}
}
\value{
a data.frame of the r-squared and the rmse for the linear regression
}
\description{
Summarize linear model fit.
}
\examples{
test <- data.frame(x = 1:100, y = 1:100 + rnorm(100))
modelfit(test, yvar = "y", xvar = "x")
}

