% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestbyday.R
\name{nearestbyday}
\alias{nearestbyday}
\title{Merge two point datasets by day.}
\usage{
nearestbyday(jointo.pts, joinfrom.pts, jointo, joinfrom, jointovarname,
  joinfromvarname, joinprefix = "closest", valuefield = "avewsp",
  knearest = 5, maxdistance = NA, nearestmean = FALSE, verbose = F)
}
\arguments{
\item{jointo.pts}{points matrix output from \code{\link{makepointsmatrix}}}

\item{joinfrom.pts}{points matrix output from \code{\link{makepointsmatrix}}}

\item{jointo}{data.table target}

\item{joinfrom}{data.table data source}

\item{jointovarname}{column name for unique field for points in jointo}

\item{joinfromvarname}{column name for unique field for points in joinfrom}

\item{joinprefix}{string to use as name for new variables}

\item{valuefield}{column name with field to be merged in}

\item{knearest}{number of nearest neighbors to check for non-missing data}

\item{maxdistance}{(optional) constraint on distance}

\item{nearestmean}{(optional) compute mean of knearest values within maxdistance?}

\item{verbose}{print output of the intermediate data.table size and fields}
}
\value{
A data.table, akin to a semi_join on nearest neighbor and day.
}
\description{
Merge point datasets by day to the closest of K nearest neighbors 
 (with an optional distance constraint).
}
\examples{
jointo <- data.table(x = 1:5, y = 1:5, siteidx = 1:5,
                    day = as.Date("2004-04-01"))
joinfrom <- data.table(x = rep(1:5 + 0.1, times = 5),
                      y = rep(1:5, each = 5), 
                      siteidy = letters[1:25], # currently this needs to be character
                      value = 100 + 1:25,
                      day = as.Date("2004-04-01"))
                      
jointo.pt <- makepointsmatrix(datatable = jointo, 
                             xvar = "x", yvar = "y", idvar = "siteidx") 
joinfrom.pt <- makepointsmatrix(datatable = joinfrom, 
                               xvar = "x", yvar = "y", idvar = "siteidy") 
joinout <- nearestbyday(jointo.pts = jointo.pt, joinfrom.pts = joinfrom.pt, 
                       jointo = jointo, joinfrom = joinfrom, 
                       jointovarname = "siteidx", joinfromvarname = "siteidy", 
                       joinprefix = "nearest", valuefield = "value", 
                       knearest = 3, maxdistance = 2, 
                       nearestmean = TRUE, verbose = T)
joinout                        
}

